+++ db/migrate/20240131082218_add_reset_sent_at_to_users.rb
@@ -1,0 +1,5 @@
+class AddResetSentAtToUsers < ActiveRecord::Migration[7.1]
+  def change
+    add_column :users, :reset_password_sent_at, :datetime
+  end
+end